+++ lang/go/cndp.org/examples/fwd/fwd.json
@@ -41,7 +41,7 @@
             "qid": 37,
             "umem": "umem0",
             "region": 1,
-            "description": "LAN 0 port"
+            "description": "LAN 1 port"
         }
     },
