+++ lang/go/cndp.org/examples/fwd/run_fwd
